[power][storage] Create system integration test
This change adds the system integration directory with a test created
for storage. This test includes a custom product and board that infra
will flash onto a vim3 device and then run the test in isolation. The
custom product is a bootstrap level product, so it will be run on
bringup builders.

Bug: 335526421
Change-Id: Ibe7a0a97548342cd826cc4be868d06adef7f18de
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1042046
Commit-Queue: Novin Changizi <novinc@google.com>
Reviewed-by: David Fang <fangism@google.com>
Fuchsia-Auto-Submit: Novin Changizi <novinc@google.com>
Reviewed-by: Didi She <didis@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>