[components] Namespaces support EXECUTABLE right
With this change, the namespace directories that are built by component
manager have the EXECUTABLE right.

This change does not effect on service nodes (protocols). Directories
that are added to the namespace (from cml `directory` capabilities for
example) are also not affected, except that they now support EXECUTABLE.

Bug: 337937584
Change-Id: Ib5d2c2041aa7031c26145fb209effd57edc615a6
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039215
Reviewed-by: Chris Suter <csuter@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
Reviewed-by: Mark Dittmer <markdittmer@google.com>
Commit-Queue: Gary Bressler <geb@google.com>