[sdk][fidl] add comment to fuchsia.hardware.power.SagElement
When searching for a list of SAG elements, I was confused by the
`SagElement` enum. I thought the enum specified all of the
SAG elements.

That turns out not to be the purpose of the `SagElement` enum.
Rather, the enum is meant to be used by clients of the power
support library. (In retrospect, `fuchsia.hardware.power`
does suggest something driver-specific.)

To help my future self, add a comment explaining the purpose
of `SagElement`.

Alternatively: we could add the currently missing element to
the enum. But that has a couple of problems:

1. It invites driver authors to use the enum value, although
   that value isn't supported by the power support library.
2. It creates a maintenance burden, if/when we add additional
   elements to SAG.

Change-Id: I9958a71ee37d6c6713da2ac9df841f6ec5d1887e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046092
Reviewed-by: Michael Brunson <mbrunson@google.com>
API-Review: Michael Brunson <mbrunson@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Fuchsia-Auto-Submit: Mukesh Agrawal <quiche@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>