[fxfs][fsck] Fix spurious fsck error
Running fsck on a filesystem with an empty volume was erroneously
failing.  The allocator reports (correctly) that 0 bytes are allocated
to the volume, but Fsck fails to filter this out and complains that the
owner_allocated_bytes is different than what is observed.

Bug: 306749272
Change-Id: I529653d72e20dbc49ca2627dda4caa0fe7d8fc19
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045220
Reviewed-by: Aaron Drew <ripper@google.com>
Fuchsia-Auto-Submit: James Sullivan <jfsulliv@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>