[netstack3] Introduce TransportPacketMut
Introduce variants of the transport packet methods/traits in the
filtering crate that allow for packet modification (not just
inspection). This includes MaybeTransportPacketMut and
TransportPacketMut. For now, the only methods on the latter are for
updating the pseudo-header in the transport-layer checksum, which will
be necessary to allow updating the source/destination IP addresses in
the IP header.

For packet builders, updating the pseudo-header is simple because the
packet has not yet been serialized and the checksum has yet to be
calculated. For incoming packets that are partially parsed, introduce
the ParsedTransportHeaderMut type, which holds a parsed transport-layer
packet allowing for in-place rewriting of its fields.

Bug: 321013529
Test: netstack3-filter-test
      netstack3-core-test
Change-Id: Ice0fdc1d069f98af4f5b717ccfa4e7e43ada8d8d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044975
Reviewed-by: Peter DeLong <pdelong@google.com>
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>