[bazel_sdk] Make fuchsia_transition() idempotent.
Due to the way Bazel works, the transition function can be
called multiple times in a chain of dependencies, and should not
change the content of the input build configuration if it is already
the result of a previous fuchsia_transition() call.

This fixes an issues where `-ffuchsia-api-level=XXX` was appended
to the `copt` list on each call, resulting in a different build
configuration instance on each call.

+ Add @fuchsia_common//:transitition_utils.bzl to export new
  functions `set_command_line_option_value()` and
  `remove_command_line_option_values()`, as well as appropriate
  unit-tests, which can be run with:

  ```
  fx bazel test @prebuilt_clang//:test_suite
  ```

+ Remove obsolete key from NATIVE_CPU_ALIASES, the
  associated value (aarch64) was wrong anyway.

Change-Id: Ib43f3666b97faa38cb059548d876938cb2b7ff04
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047975
Reviewed-by: Chase Latta <chaselatta@google.com>
Commit-Queue: David Turner <digit@google.com>