[netstack3] Prepare loopback for crate split
Split everything that will be in the integration core crate into a
separate file. We had to mark tx and rx queue metadata separately to
satisfy the fact that we need different OrderedLockAccess
implementations for the tx and rx queue state. The pattern will need to
be applied again to Ethernet and IP devices in a follow up CL.

This is still not fully clean, the blanket implementation of Receive and
Transmit queues BindingsContext traits will not work post crate split.
We'll need to massage the traits there.

Bug: 338448926
Change-Id: Ie8e6b50539d4e9f84a0ebe45f3f92ec4fae0b6bf
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048414
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>