[fx] Updating remote scripts to use ~/ssh keys
This removes the consideration of the in-tree ssh key when checking
ssh keys between the local and remote computers.

This is being done to align in-tree and out of tree ssh keys by using
`ffx config get ssh.(priv|pub)` as the source of truth for the location
of the keys both in and out of tree.

Bug: b/339034335
Change-Id: I90a1e4c7bf688af9a6e26e1ca74baa379c433d03
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045213
Commit-Queue: Clayton Wilkinson <wilkinsonclay@google.com>
Reviewed-by: Colin Nelson <colnnelson@google.com>