[netstack3] Refactor IP dispatch to use packets
... rather than raw buffers, which enables a follow up CL to deliver the
packet to raw IP sockets.

Because `Ipv{4,6}Packet` contains a reference to the underlying buffer,
the packet is invalidated whenever the underlying buffer is moved (due
to Rust's move semantics). This happens when passing the buffer as an
argument to a function. By passing the packet rather than the buffer,
the packet remains valid, and can be used both for raw IP socket
delivery and delivery to the transport layer.

Note, that one drawback of this approach is that the transport layer no
longer takes ownership over the underlying buffer. This may have been
useful for optimizations that want to reuse the backing allocation, but
such optimizations are too far down the road to realistically plan
for.

Bug: 337825157
Change-Id: I692ea45e5b1029a9bc3b7f01ce5389d87fb2ed96
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1048352
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>