[audio][drivers] move to started/bypassed
Migrate drivers and driver clients away from signalprocessing
property `enabled`, to `started` and/or `bypassed`.

Field enabled is already in the process of being deprecated.
Also a coming CL creates a subset of ElementState that will
be used for SetElementState, and it does not include enabled,
so this CL is early preparation for that change.

Change-Id: Ie64fdade222b1e71aecd3a23b921f0d584b14d5c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046066
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>