[display][coordinator] Remove flaky pattern from IntegrationTest.
The Display Coordinator integration tests use an anti-pattern that has
consistently proven to result in flaky tests. The anti-pattern is using
a polling mechanism to check whether a different thread has completed
some work, combined with a timeout. The timeout is an inevitable source
of flakiness due to false negatives (tests failing despite the
implementation being correct), because tests are typically ran on shared
infrastructure ("cloud") with no worst-case performance guarantees.

Fortunately, most instances of this anti-pattern use the timeout as an
early failure mechanism -- the test is considered to have failed if a
predicate does not become true within the timeout. This CL removes the
timeout altogether in these instances. The test framework's timeout
mechanism will flag the test failures that were previously caught by the
polling timeout.

There are a couple of instances where the anti-pattern is used in the
opposite way, where tests verify that a certain condition does not
become true within some amount of time. This usage can introduce
flakiness as well, due to false positives (tests passing despite an
incorrect implementation), because the timeout may occur before the
implementation performs the undesired action. Fixing this usage is
deferred, and this CL adds clear comments where this is the case.
Deferring is considered acceptable because a correct implementation will
not exhibit the test flakiness.

This CL reimplements the polling mechanism described above. The new
implementation uses patterns that are easier to reason about. Manual
lifecycle management is replaced by smart pointers. Atomics and C-style
completion signals are replaced by a mutex and a condition variable.

This CL disables IntegrationTest.EmptyConfigIsNotApplied, which now
fails somewhat consistently. Re-enabling will happen in a follow-up.

Bug: 338379402
Test: fx test display-coordinator-unittests
Change-Id: I8e9842cc951cfb77f4879a95571a1288c4f06f18
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/884090
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Yilong Li <liyl@google.com>