[display][coordinator] Do not keep fdf::Arena in EngineDriverClient.
fdf::Arena and fidl::Arena are designed to be for ephemeral use only.
Once a piece of memory is allocated in an Arena, it's never recycled
until the Arena is destroyed.

Thus, the fdf::Arena should never be kept as a member variable in
display coordinator EngineDriverClient, otherwise it will eventually
cause an OOM.

This change replaces the arena_ member variable with local fdf::Arena
created within the FIDL call scope; this is the common practice in
driver code across the Fuchsia repository. Note that each fdf::Arena
creation calls the fdf_arena_create() function, which will allocate a
new Arena on the heap.

Test: "display-tool squares" on core.vim3
Bug: 325474586
Change-Id: I2c23191f190ab5ee61b843a9e2c8e90bef24f7a0
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046639
Commit-Queue: Victor Costan <costan@google.com>
Reviewed-by: Caroline Liu <carolineliu@google.com>
Reviewed-by: Victor Costan <costan@google.com>
Fuchsia-Auto-Submit: Victor Costan <costan@google.com>