[audio][drivers] Add ProcessingDelay property
Add `processing_delay` to ElementState and deprecate `latency`.
Although the `latency` union is a superset of processing_delay,
only the duration aspect of `latency` is ergonomically optimal
(the other option is to specify this in frames), so in practice
it is simpler to just be a duration, named ..._delay like the
others (external_delay, turn_on_delay, etc).

Change-Id: Id6f511c00aa29e70f15a60f84d87c4db3cda0bc8
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046071
API-Review: Yegor Pomortsev <ypomortsev@google.com>
Commit-Queue: Martin Puryear <mpuryear@google.com>
Reviewed-by: Yegor Pomortsev <ypomortsev@google.com>