[ctf] soft migrate function button assertion
The function button currently maps to a power media buttons event. We
need to soft migrate the CTF tests on the media button listeners to stop
asserting this is the case. Once this change rolls, we can add function
button behavior to map to a function media buttons event, and reinstate
a new assertion accordingly.

Bug: 329271369
Change-Id: Ie65f1471692bda58bc06473bb4257879b9c2812b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049612
Reviewed-by: Jianpeng Chao <chaopeng@google.com>
Commit-Queue: Caroline Liu <carolineliu@google.com>