[component_manager] pass program output into dictionary construction
Passes the program output dictionary directly into dictionary
construction instead of reaching into the component for it. This is
necessary because we won't be able to access component internals so
easily once this code is moved to the routing library.

This change would have introduced reference cycles between dictionaries,
so a WeakDict type is introduced. A WeakDict can be upgraded into a
Dict. This was not an issue before because all dictionaries that wanted
to route into a program output dictionary would hold a weak reference to
the component, and upon receiving a routing request they would then
upgrade that reference and grab the component's program output
dictionary from there.

Change-Id: I1b182819326abff95150eb0e9b0585f514c9221d
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044832
Commit-Queue: Gary Bressler <geb@google.com>
Reviewed-by: Gary Bressler <geb@google.com>