[superproject] Roll scudo [scudo] Move the chunk update into functions (#83493)
The code paths for mte enabled and disabled were interleaving and which
increases the difficulty of reading each path in both source level and
assembly level. In this change, we move the parts that they have
different logic into functions and minor refactors on the code
structure.

GitOrigin-RevId: ce5d0c108782f917253af23a31f16410887f5447
Original-Revision: ed22bf262875e3826259f5b65d9825c80b4cec84
Roller-URL: https://ci.chromium.org/b/8747742043800232065
CQ-Do-Not-Cancel-Tryjobs: true
Change-Id: Ie545eee4bbea2c0570c61d69e6fa352dc60ad803
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049727