[starnix] Handle state explosion due to branches
The previous verification algorithm will explore each possible branch of
the program. This induced an algorithm that was exponential in the
number of conditional branches.

To solve this issue, this CL introduces an algorithm inspired by the
description of the linux kernel algorithm in https://docs.kernel.org/bpf/verifier.html

The verifier use liveness information to broaden any state it can verify
will correctly terminate. It then keep track of all these states and
when a new state is produced, it checks whether it can already know it
will terminate, and in that case, skip the computation.

This CL also adds a test that exhibit the exponential behaviour of the
previous implementation.

Change-Id: I18d703ca896f24c3d04af0e751ba218732672ad1
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1039632
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Reviewed-by: Kevin Lindkvist <lindkvist@google.com>
Fuchsia-Auto-Submit: Benjamin Lerman <qsr@google.com>