[fuchsia-async] Add Scope structured concurrency API
The goal of this API is to improve reasoning about task lifetimes and
program behavior.  For now it is considered a prototype and is
`#[doc(hidden)]` to allow us to iterate freely.

The Scope owns its own map of tasks and wakers, giving us more granular
locking and task management. This is based on an earlier prototype by
csuter. Everything tied to the executor main loop is still global,
including the ready queue.

The drop behavior of collecting tasks onto the heap also means we can
remove the atomic check before spawning a task to prevent deadlocks in
edge cases where a task is spawned from a destructor during shutdown
(this already has a test).

Bug: 339724492
Fixed: 332915575
Change-Id: I90658c95753739079897ccb3f17dfae1f774d684
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1040152
Reviewed-by: Chris Suter <csuter@google.com>
Commit-Queue: Tyler Mandry <tmandry@google.com>
Size-Review: James Sullivan <jfsulliv@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>