[kernel][fasttime] Implement libfasttime
Adds a new kernel library called libfasttime that allows callers to
retrieve ticks and compute CLOCK_MONOTONIC without having to make a
a syscall on platforms that support it. This library reads values
associated with time, such as the raw_ticks_to_ticks_offset and the
ratio between ticks and monotonic time, out of a new, read-only VMO
called the TimeValuesVmo. As implemented, this VMO is actually a child
slice of the vDSO VMO, which allows the time constants to be shared
between syscalls, vDSO functions, and libfasttime. A handle to this VMO
is provided to userboot, which then makes it available to userspace.

This library will be used by starnix to compute CLOCK_MONOTONIC from the
Linux vDSO, which does not have access to clock syscalls.

Bug: 328306129
Change-Id: I2d6ebb3838997e400715de04cdecd972469648ad
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1045493
Reviewed-by: John Grossman <johngro@google.com>
Fuchsia-Auto-Submit: Anirudh Mathukumilli <rudymathu@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>