[ld] Support preloaded modules in RuntimeDynamicLinker::Init
This allows an InitModule to indicate that a module should be
consulted for dynamic linking but is already present in the
address space and does not need to be loaded.

The new ld::RuntimeDynamicLinker::PreloadedImplicit convenience
function turns the initial modules from one dynamic linking
session into preloaded implicit modules for a new session.

This is demonstrated in a test where a main executable with a
DT_SONAME and dependencies of its own is launched via remote
dynamic linking, and then later a second root module is loaded
into its own dynamic linking namespace using the main executable
and its vDSO as preloaded implicit modules to satisfy a DT_NEEDED
on that SONAME.  This approximates the intended case of a driver
host process and a separate driver loaded into it by a service.

Bug: 326524302
Change-Id: I900055e832d04d20f8f1cbb058856db07c462459
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047433
Reviewed-by: Caslyn Tonelli <caslyn@google.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>