[fidl][fuchsia.diagnostics] platform-only APIs
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or external components as such.

LogSettings is not allowed to have clients or servers outside of the
platform.

ArchiveAccessor can have clients outside the platform but servers can
only be in the platform.

Bug: 335446415
Change-Id: Ifa7880d2293065e784943b3f4bc549e84061c44e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1049676
API-Review: Miguel Flores <miguelfrde@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Ian McKellar <ianloic@google.com>
Reviewed-by: Miguel Flores <miguelfrde@google.com>