[fidl][kernel] Mark protocols as server=platform
As part of RFC-0241 we are marking protocols whose servers or clients
are only supposed to be in platform or extenal components as such.

This marks all of the fuchsia.kernel protocols as allowing clients in
platform or external components but only allowing servers in platform
components.

Bug: 335446415
Change-Id: I51e00847e26cb4337d57f5d41f67e5919db12c02
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046054
Commit-Queue: Ian McKellar <ianloic@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
API-Review: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Olivia Nordquist <nolivia@google.com>