[netstack3/conntrack] Fix self-connected sockets
Conntrack has tacitly assumed that for all connections, the original and
reply tuples will be different, but this is not the case with
self-connected sockets.

There are a couple of fixes required here. The primary one is to track
the number of connections separately from the HashMap. Then, we also
need to insert a few ifs to fence off a couple asserts.

Change-Id: Ic228837e1087afdc3d4f58cc5d6b748de1be1e5a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050513
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>