[netstack3] Reduce freestanding fns in socket mod
All the helpers in the netstack3_core::socket module will be moved to
the base crate. Tying the freestanding functions to well named types
makes them more ergonomic in the long run and sidesteps the desire to
import freestanding fns.

While at it, noticed that the TCP module was attempting to upgrade a
bound device ID in an error path which should not be done so massage the
helpers in socket and TCP to correct that.

Bug: 338448926
Change-Id: I17333ad6a5fcbd1be01a6d3dcfb17c22cdf5b2e3
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050598
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Bruno Dal Bo <brunodalbo@google.com>