[wlan] Handle framework errors in fidl_bridge
ForwardResult now explicitly checks for framework errors. ForwardResult
will now close the completer if the framework error is due to channel
closure.

For errors that do not result in channel closure, ForwardResult sends
the framework error to the caller as a domain error if possible.
Otherwise, it logs and ignores the error and replies to the caller with
a default-initialized response if necessary.

Unit tests have been added to show that channel closure propagates
through a bridge server using ForwardResult.

Test: fx test drivers-only/fullmac fidl_bridge_test
Change-Id: Iab0b632f14734214c311c165674ace0074f52050
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1043192
Reviewed-by: Charles Celerier <chcl@google.com>
Commit-Queue: John Tan <johntan@google.com>