[netstack3] Hook up conntrack
There's nothing that depends on conntrack yet, but hooking it up means
we'll get a lot more coverage to help uncover any issues.

Note that there are no paths currently that drop a packet due to
conntrack.  This will change when NAT is integrated, but for now it
means that we'll have less impact on other workloads if there are
problems.

Hooking this up, revealed a few tests that were waiting for the netstack
to reach quiescence, or that were depending on an exact set of timers
being installed, which the GC timer throws a wrench into. To avoid these
issues, filter out any attempts to schedule the GC timer in Core tests.

Additionally, return None from MaybeTransportPacket and
MaybeTransportPacketMut for InnerSerializer<&[u8], EmptyBuf>, which is
now load-bearing. Also remove the impls for Buf<Vec<u8>> after moving
the since test relying on it to use InnerSerializer<&[u8], EmptyBuf>
instead.

Fixed: 340854102
Change-Id: I443ce8e712eb2784a2578182767a65397ec5ced4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044954
Reviewed-by: Peter Johnston <peterjohnston@google.com>
Commit-Queue: Peter DeLong <pdelong@google.com>