[netstack3] Add IpPacket::set_{src,dst}_addr
Add methods to IpPacket that allow modifying the source and destination
IP addresses. For IPv4 packets, this requires updating the checksum, and
for all IP packets that contain a valid transport header, it involves
updating the transport-layer checksum to reflect the new address in the
pseudo-header.

For ForwardedPacket, which wraps an entire IP packet which has had the
IP header parsed to allow for buffer reuse, updating an IP address in
the packet header involves reparsing the IP header so it can be modified
in place, and then updating the checksum(s) as necessary.

Bug: 321013529
Test: netstack3-filter-test
      netstack3-core-test
Change-Id: I5188edf90692b7b3692ab56fdf88baf5669bd6ee
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1044953
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Peter Johnston <peterjohnston@google.com>
Fuchsia-Auto-Submit: Peter Johnston <peterjohnston@google.com>
Reviewed-by: Peter DeLong <pdelong@google.com>