[starnix] Introduce framebuffer2 feature
This feature allows views to be associated with containers, instead of
components that run inside of containers. In the future the framebuffer
feature will be replaced with framebuffer2, but that requires a soft
transition in other repositories.

Bug: b/337303619
Change-Id: I8c622ae43228da00fd56aa5c94942297e335aa5e
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050253
Reviewed-by: Emircan Uysaler <emircan@google.com>
Fuchsia-Auto-Submit: Kevin Lindkvist <lindkvist@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>