[bazel_sdk] Fix structured configs with 'config' section in CML file
The GN version of fuchsia_component_manifest has a hardcoded bit
of code that references structured configs. This copies that logic
over to the bazel template: http://shortn/_DWGNHmiTHD

Also fixed an issue that caused the outer brackets of the values dict
to be dropped when it was written to a file.

This also allows boolean values to be written as boolean values
directly, not strings of "true" or "false".

Change-Id: I2be528cd55f535e94b93427e2c4bed5c6288cd01
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052052
Reviewed-by: Jiaming Li <lijiaming@google.com>
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Jerry Belton <jcecil@google.com>