[remote-control] Remove ephemeral resolution of children on userdebug
The ffx-laboratory collection was using laboratory-env, which allows
ephemeral resolution on userdebug (i.e. resolution from remote package
repositories instead of only from base packages). Ephemeral resolution
is generally prevented on userdebug and I suspect that this use was
allowed because of difficulties in configuring the system which have
since been ameliorated.

This CL switches ffx-laboratory from laboratory-env to core-env, which
differs only in that it does not allow ephemeral resolution on
userdebug, i.e. core-env only allows resolution of base packages on
userdebug. It also deletes laboratory-env as this was the last use.

Children of the ffx-laboratory collection will now need to be in base
on userdebug builds.

Bug: 341179052
Change-Id: Iddbc2d93e93aabe924674f835073f4f9cce4f2a4
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050566
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Aidan Wolter <awolter@google.com>
Reviewed-by: Casey Dahlin <sadmac@google.com>
Owners-Override: James Robinson <jamesr@google.com>
Commit-Queue: Ben Keller <galbanum@google.com>