[sysmem2] better defaults when secure_required
Avoid creating an inconsistency (and constraints validation failure)
via cpu_domain_supported and/or ram_domain_supported defaults when
those fields aren't set and secure_required is set to true.

Also log usage bits when SetVerboseLogging, and try to keep the old
client name when the client name is set, as long as the length doesn't
get too excessive, for easier tracking of which Node is which.

BUG: b/306258175
Change-Id: I49a6862c13bbddce40ce0babfc1c7481f50dad88
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051361
Reviewed-by: John Bauman <jbauman@google.com>
Commit-Queue: Dustin Green <dustingreen@google.com>