[hrtimer][vim3] Fix lease management
jmatt@ noticed that this driver was not managing the power element
in order to get the lease satisfied. In this CL we add RequiredLevel
and CurrentLevel support to get the state of the level updated, and
we monitor the lease status before considering the lease granted.

Previous testing of lease management did not reveal this problem
because an unsatisfied lease also prevents suspension but does not
guarantees it as is the intent of this driver.

Fixed: 341701622
Test: fx test -o aml-hrtimer-test
Test: hrtimer-ctl --id 0 --keep_alive 30 --wait 0 prevents suspension
      triggerd via adb shell "echo mem > /sys/power/state" for 30 seconds
      (works without this change, checking that it still works)
Change-Id: Ib9514bb18b217830afa24f59e1ae77926a625b24
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051034
Reviewed-by: Kyle Gong <kgong@google.com>
Reviewed-by: Zhiyi Chen <zhiyichen@google.com>
Reviewed-by: Justin Mattson <jmatt@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>