[kernel][gdb] Make one attempt to locate the kernel symbols.
When attempting to determine where the kernel has been relocated to
when loading symbols, make one last try to find the relocation address
using the default offset of 0x10000000 before giving up.

This script is not in great shape right now, but while it has a large
number of imperfections, it can still be very useful when attempting
to pick apart the state of a dead QEMU based system.  This hack can
keep it a bit more useful until it can (someday) be replaced with
something better.

Change-Id: I8710285d179534cbfccab4eef2116fa962351a4a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053753
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Commit-Queue: John Grossman <johngro@google.com>