[fuchsia.io] Remove Enumerate/DirectoryIterator
To speed up io2 migration, remove Enumerate (and DirectoryIterator) from
the protocol.

Enumerate/DirectoryIterator would need to be implemented as part of the
Open2 migration. There is no current server support for it and there
does not seem to be a pressing need to implement it at the moment. Most
cases are still served by ReadDirents. Removing Enumerate is to help
simplify the io2 migration process, it can be added back when migration
is completed.

Bug: b/340436161
Change-Id: I4003fb0db9d48276d796a75d2a40cddd8ffdc9ca
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1047692
Reviewed-by: Sen Jiang <senj@google.com>
Reviewed-by: Gary Bressler <geb@google.com>
Reviewed-by: Brandon Castellano <bcastell@google.com>
API-Review: Adam Barth <abarth@google.com>
Reviewed-by: Adam Barth <abarth@google.com>
Commit-Queue: Rui Qi Sim <rqsim@google.com>