[netstack3] Avoid 'a lifetime in context_ip_bounds
If the context_ip_bounds is used to annotate an implementation that
already has a 'a lifetime in scope, the lifetimes collide and lead to
some obscure compiler errors. Avoid the confusion by giving the macro a
dedicated lifetime name that's less likely to collide with anything
declared elsewhere.

Change-Id: I8f5426263a7da70fcf3fac3103c8df859637cd37
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053574
Reviewed-by: Bruno Dal Bo <brunodalbo@google.com>
Commit-Queue: Jeff Martin <martinjeffrey@google.com>