[elfldltl][phdr] Refactor _dl_locked_report_globals
This is a first step in the work to disentangle libc code from
the musl dynamic linker internals.  _dl_locked_report_globals in
the musl implementation is refactored to use a new C++ subroutine
that will be reused in the integration of the new dynamic lniker.

Bug: 342469121
Change-Id: Iaa3ee57e1daf6506c2ea7ee25499361e10b5970b
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054072
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>
Fuchsia-Auto-Submit: Roland McGrath <mcgrathr@google.com>
Reviewed-by: Caslyn Tonelli <caslyn@google.com>