[paver] Guards flaky test with block watcher pauser
The OnStop* tests are occasionally flaky and report that fs watcher unexpectedly sees block device disconnected in the middle of getting partition info. This could be caused by InitPartitionTables rebinding the gpt driver. The tests don't require fs watcher to be running thus we disable it for the entire test to see if the flake gets better.

Bug: 317584846
Change-Id: Ib6896ebbfb7a72e6f89388974a21f6e882f3b4ff
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053429
Reviewed-by: David Pursell <dpursell@google.com>
Commit-Queue: Yecheng Zhao <zyecheng@google.com>