[kernel][handle] Make observer's handle a void ptr
In the signal observer machinery the handle (class Handle) is
only used as a key. In other words the only thing that matters
is its address, to match for cancellation.

While having the type explicit is helpful as documentation
it gets in the way for evolving the handle table implementation
by tying the observer machinery with the handle table types
which is not required.

TEST=current testing suffices
BUG=fxbug.dev/42077761

Change-Id: I01ee8be28742adf2558908e6f3a3ac3cf506d61c
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1053515
Commit-Queue: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: Nick Maniscalco <maniscalco@google.com>
Reviewed-by: Anirudh Mathukumilli <rudymathu@google.com>