[netstack3] Split IP module into its own crate
This change extracts the netstack3 IP module into the netstack3_ip
crate. The new crate follows the pattern of re-exporting everything from
the core of the library. A flatter organization is preferred there, with
a root re-export module and a small number of very flat modules. Maximum
depth is 3 for testutils only. A handful of symbols were renamed to
allow unambiguous re-export.

Most of the changes are making symbols `pub` and adding documentation,
so we can fulfill disallowing missing_docs. Things that should only be
used in tests are marked as such, but many structs that had `pub(crate)`
field access were just promoted to `pub`. `#[allow(missing_docs)]` is
used in a handful of places, mostly in self-explanatory enum variants.

Note that the datagram module had to be extracted into the IP crate as
well to satisfy the local ICMP echo socket implementation. The plan is
to extract it from netstack3_ip later on.

Notable non-obvious changes:
* All the uninstantiable implementations are moved from core into
  `uninstantiable` mods.
* The NUD delegation trait changed its shape a bit to avoid a big rework
  in the ARP module's tests, which hit coherence errors.
* `FilterHandlerProvider`, which was defined in an integration module,
  was lifted to netstack3_ip::base.
* The large implementations on common test fakes has been moved from
  core to IP as well.

For now, core is re-exporting all of `netstack3_ip` where its `ip`
module was, and using some tricks to get all the files compiling. Follow
up changes will extract integration tests closer to the IP crate,
re-organize the integration files with better names, and stop glob
re-exporting from the IP crate.

Bug: 338449006

Change-Id: I4795909efbb992bf158b079b0d99940dc3e226bc
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1051359
Reviewed-by: Jeff Martin <martinjeffrey@google.com>
Commit-Queue: Bruno Dal Bo <brunodalbo@google.com>