[async-loop] Switch to zx_port_cancel_key on Fuchsia >= 20
This updates the async-loop library to use the new zx_port_cancel_key()
API when building for API level 20 or above.

This means that the async_wait structure's `object` field is no longer
used by the async loop library. Many users of this library use this
field to store data for other uses, so the field is left alone. We may
consider removing this field in the future to reduce the size of this
structure in memory.

Change-Id: I60d668749b74286cc1b7a707a34ce8759f5c3bcd
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1046693
Reviewed-by: David Dorwin <ddorwin@google.com>
Commit-Queue: James Robinson <jamesr@google.com>
Reviewed-by: Carlos Pizano Uribe <cpu@google.com>
Reviewed-by: John Bauman <jbauman@google.com>