[hrtimer][vim3] Re-enable unit tests
The previous use of fidl::ServerBindingRef in the unit testings
allowed for potentially calling the CloseHandler method (that
captures the this pointer to access the member variable
lease_requested_) after the enclosing object went out of scope.
The use of fidl::ServerBindingRef was removed in fxrev.dev/1051034
in favor of fidl::ServerBinding which does not have this problem.

Fixed: 332644994
Multiply: aml-hrtimer-test
Change-Id: I10faca24bc5a41d535970af86a8c4d887bb3a770
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1050602
Reviewed-by: HanBin Yoon <hanbinyoon@google.com>
Commit-Queue: Andres Oportus <andresoportus@google.com>