[kernel][memalloc] Ensure that a failed insertion does not modify the pool
There was a corner case within the "insert subrange" routine that
modified internal state ahead of being sure that it had sufficient nodes
to successfully follow through. Unsuccessful operations should not
modify the pool.

This bug was surfaced in a fuzz case that unsuccessfully resized a range
(implicitly using the routine in question) and then tried to resize it
again, which hit an assertion indicating that the was no longer tracked.

Fixed: 42080425
Change-Id: I6dd4aff57e9122b639eea222d8604bfbb5d6898a
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1056584
Fuchsia-Auto-Submit: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Auto-Submit <auto-submit@fuchsia-infra.iam.gserviceaccount.com>