[zircon] Introduce next gated syscall to manage power level transitions
This is prototyping a syscall using a port transport to allow
information flow from the kernel to userspace and from userspace back to
the kernel.

Additionally this introduces a new topic for set_performance_info which
communicates the current power level for a given power domain. The power
level matches that provided by the scheduler when requesting a
transition with an opaque argument.

In order to keep port packet relatively small (within the 32 bytes) the
cpu mask may not be used in this environment (64 bytes). To address this
the prior syscall that sets a power domain for a given cpu mask also
provides an ID, as a way to refer to all CPU that are members of the
domain.

Change-Id: Id33e6ee60a6da3d8e7475147fa4f22047ba04001
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1011960
Reviewed-by: Joshua Seaton <joshuaseaton@google.com>
Reviewed-by: Corey Tabaka <eieio@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>