[kernel] Support IRQ configuration for UART
First we introduce a new field into the zbi_dcfg_simple_t, which
repurposes an existent reserved field as a flags field from now on.

4 bits are introduced which indicate whether is level or edge triggered.
The following two bits indicate whether its active low or high, or
whether is active low-to-high or high-tolow.

Intentionally the flag field left as zero implied no configuration is
available, and will be left to whatever the platform does by default.

In some cases like RISCV, the devicetree does not provide this
information.

As of this CL, the uart IRQ information is obtained from the devicetree,
and when available is applied before registering the irq handler.

Change-Id: Ic7934a7baf2d35eabf10d1fa4a0477eb56b22a69
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1052916
Commit-Queue: Gianfranco Valentino <gevalentino@google.com>
Reviewed-by: Roland McGrath <mcgrathr@google.com>
API-Review: Roland McGrath <mcgrathr@google.com>