[kernel][vm] Synchronize reclamation with aging
Adds a mechanism for reclamation, via the PeekReclaim method, to
synchronize with any pending aging events. This means that if there is
aging that should happen, PeekReclaim will wait for it to complete
before insisting there are no pages available.
This can resolve a scenario where we falsely OOM on a heavily loaded
system.

Supporting this synchronization method required the timeout logic that
was implicitly in the MruThread to be made available to generally query
if aging should happen. To make this simpler the aging_event_ was
re-defined to what it is actually used for, just indicating an active
ratio trigger, and methods and call sites around this trigger were
cleaned up to acknowledge this specialization.

Run-All-Tests: True

Change-Id: I5fbbb4d5686a84d941e346645be57955683c8db5
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1031012
Reviewed-by: David Worsham <dworsham@google.com>
Commit-Queue: Adrian Danis <adanis@google.com>