[sensors] Playback component lifecycle changes.
Multiple connections to the Driver and Playback protocols are not
permitted. The second connection will be closed with the epitaph
ZX_ERR_ALREADY_BOUND.

When the Playback client disconnects, the playback configuration is
cleared.

When the Playback client disconnects, the Driver client will be
disconnected with epitaph ZX_ERR_BAD_STATE if present.

The Driver client may connect first, however if any API calls are
made before a playback configuration is set, the Driver client
will be disconnected with epitaph ZX_ERR_BAD_STATE.

Change-Id: I47df8c146953ddc0df72489c14868d5345b3e725
Reviewed-on: https://fuchsia-review.googlesource.com/c/fuchsia/+/1054092
Reviewed-by: John Shamoon <johnshamoon@google.com>
Commit-Queue: Matt Thiffault <mthiffault@google.com>